+++ quint/src/ToIrListener.ts
@@ -358,12 +358,15 @@ export class ToIrListener implements QuintListener {
   }
 
   // module Foo = Proto(x = a, y = b)
+  // module Foo = Proto(x = a, y = b) from "<path>"
   exitInstanceMod(ctx: p.InstanceModContext) {
     const protoName = ctx.moduleName().text
     const qualifiedName = ctx.qualifiedName()?.text
     const names = ctx.name()!
     const nexprs = ctx.expr().length
     const overrides: [QuintLambdaParameter, QuintEx][] = []
+    // slice <path> from the quoted string "<path>", if the path is present
+    const fromSource = ctx.fromSource() ? ctx.fromSource()!.text.slice(1, -1) : undefined
     if (nexprs > 0) {
       const exprs = popMany(this.exprStack, nexprs)
       for (let i = 0; i < nexprs; i++) {
@@ -385,6 +388,7 @@ export class ToIrListener implements QuintListener {
       protoName,
       overrides,
       identityOverride,
+      fromSource,
     }
     this.definitionStack.push(instance)
   }