+++ CHANGELOG.md
@@ -11,8 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Pretty-printing of expressions in REPL (#870)
 - Spread syntax for record updates (#880)
-- Save ITF traces in `quint test` (#884)
+- Save ITF traces in `quint test` (#884, #890)
 - Initial integration with the Apalache server (#871)
+- Support `import... from` for instances (#899)
 
 ### Changed
 ### Deprecated